requirements.txt
numpy
pillow
tk
scipy
